+++ docker-compose.yml
@@ -6,6 +6,9 @@ services:
     ports:
       - '5432:5432'
     restart: always
+    environment:
+      POSTGRES_HOST_AUTH_METHOD : "trust"
+
     volumes:
       - postgres-db:/var/lib/postgresql/data
   pgadmin:
@@ -31,6 +34,7 @@ services:
       POSTGRES_DB: "postgres"
       POSTGRES_HOST : "postgres"
       POSTGRES_HOSTNAME : "postgres"
+      POSTGRES_HOST_AUTH_METHOD : "trust"
 volumes:
   postgres-db:
